Fixed API

+++ src/main/java/me/arthed/walljump/utils/VelocityUtils.java
@@ -5,7 +5,7 @@ import org.bukkit.util.Vector;
 
 public class VelocityUtils {
 
-    public static void pushPlayerInFront(Player player, float horizontalPower, float verticalPower) {
+    public static void pushPlayerInFront(Player player, double horizontalPower, double verticalPower) {
         if(BukkitUtils.isVersionBefore(BukkitUtils.Version.V1_8))
             verticalPower += 0.1;
         Vector velocity = player.getLocation().getDirection().normalize().multiply(horizontalPower);
